Moved current scanner out of main
